requirements.txt
flask==2.3.3
easyocr==1.7.1
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
pdf2image==1.16.3 